pyproject.toml
[project]
name = "synqit-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aio-pika>=9.5.5",
    "aiohttp>=3.11.18",
    "cassandra-driver>=3.29.2",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "orjson>=3.10.18",
    "pika>=1.3.2",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "python-socketio>=5.13.0",
    "redis[asyncio]>=6.0.0",
    "rich>=14.0.0",
    "uvicorn>=0.34.2",
    "yt-dlp>=2025.4.30",
]
